caf: Move key_id.h file to new include directory
Change moves key_id.h file to new include directory.
This is done for consistency with other CAF headers.
