+++ applications/nrf_desktop/configuration/nrf52kbd_nrf52832/fn_keys_def.h
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
  */
 
-#include "key_id.h"
+#include <caf/key_id.h>
 
 /* This configuration file is included only once from fn_keys module and holds
  * information about key with alternate functions.
