+++ subsys/caf/CMakeLists.txt
@@ -4,10 +4,5 @@
 # SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
 #
 
-zephyr_include_directories(
-	events
-	include
-)
-
 add_subdirectory(events)
 add_subdirectory(modules)
